refactor: extract attack log helpers
